requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
banks==2.1.1
beautifulsoup4==4.13.3
blake3==1.0.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.9.2
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
greenlet==3.1.1
griffe==1.7.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
interegular==0.3.3
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.23
langchain-core==0.3.51
langchain-ollama==0.3.1
langchain-text-splitters==0.3.8
langgraph==0.3.27
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.27
lark==1.2.2
llama-cloud==0.1.18
llama-cloud-services==0.6.9
llama-index==0.12.29
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.29
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-huggingface==0.5.0
llama-index-llms-huggingface-api==0.4.1
llama-index-llms-ollama==0.5.4
llama-index-llms-openai==0.3.32
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
llguidance==0.7.13
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
numpy==2.0.2
ollama==0.4.7
openai==1.72.0
opencv-python-headless==4.11.0.86
orjson==3.10.16
ormsgpack==1.9.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.13.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
striprtf==0.0.26
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
langchain-openai==0.0.8
